Verify the message-too-large read error
This is a critical part of the read API.  Because the default message
size write limit (1mb) is larger than the default read size (8kb),
clients using the default limits can write messages they cannot read.
Clients needs to be able to distinguish this message-too-big-to-read
case from other errors so they can retry the read with a larger ReadLimit.